+++ tests/trustless_gateway_car_test.go
@@ -30,8 +30,7 @@ func TestTrustlessCarPathing(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirTwoSingleBlockFilesFixture.MustGetCid("subdir", "ascii.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							subdirTwoSingleBlockFilesFixture.MustGetCid(),
 							subdirTwoSingleBlockFilesFixture.MustGetCid("subdir"),
@@ -55,8 +54,7 @@ func TestTrustlessCarPathing(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(singleLayerHamtMultiBlockFilesFixture.MustGetCid("685.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(flattenStrings(t,
 							singleLayerHamtMultiBlockFilesFixture.MustGetCid(),
 							singleLayerHamtMultiBlockFilesFixture.MustGetCIDsInHAMTTraversal(nil, "685.txt"),
@@ -81,8 +79,7 @@ func TestTrustlessCarPathing(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(dirWithDagCborWithLinksFixture.MustGetCid("document", "files", "single")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(flattenStrings(t,
 							dirWithDagCborWithLinksFixture.MustGetCid("document"),
 							dirWithDagCborWithLinksFixture.MustGetCid("document", "files", "single"),
@@ -115,8 +112,7 @@ func TestTrustlessCarDagScopeBlock(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirTwoSingleBlockFilesFixture.MustGetCid("subdir")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							subdirTwoSingleBlockFilesFixture.MustGetCid(),
 							subdirTwoSingleBlockFilesFixture.MustGetCid("subdir"),
@@ -139,8 +135,7 @@ func TestTrustlessCarDagScopeBlock(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirTwoSingleBlockFilesFixture.MustGetCid("subdir", "ascii.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							subdirTwoSingleBlockFilesFixture.MustGetCid(),
 							subdirTwoSingleBlockFilesFixture.MustGetCid("subdir"),
@@ -166,8 +161,7 @@ func TestTrustlessCarDagScopeBlock(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(singleLayerHamtMultiBlockFilesFixture.MustGetCid("1.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(flattenStrings(t,
 							singleLayerHamtMultiBlockFilesFixture.MustGetCid(),
 							singleLayerHamtMultiBlockFilesFixture.MustGetCIDsInHAMTTraversal(nil, "1.txt"),
@@ -203,8 +197,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirTwoSingleBlockFilesFixture.MustGetCid()).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							subdirTwoSingleBlockFilesFixture.MustGetCid(),
 						).
@@ -226,8 +219,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(singleLayerHamtMultiBlockFilesFixture.MustGetCid()).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								singleLayerHamtMultiBlockFilesFixture.MustGetCid(),
@@ -251,8 +243,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "ascii.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid(),
@@ -278,8 +269,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid(),
@@ -306,8 +296,7 @@ func TestTrustlessCarDagScopeEntity(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(dirWithDagCborWithLinksFixture.MustGetCid("document")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(flattenStrings(t,
 							dirWithDagCborWithLinksFixture.MustGetCid(),
 							dirWithDagCborWithLinksFixture.MustGetCid("document"),
@@ -339,8 +328,7 @@ func TestTrustlessCarDagScopeAll(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(
 								t,
@@ -367,8 +355,7 @@ func TestTrustlessCarDagScopeAll(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid(),
@@ -454,8 +441,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(flattenStrings(t,
 							subdirWithMixedBlockFiles.MustGetCid(),
 							subdirWithMixedBlockFiles.MustGetCid("subdir"),
@@ -482,8 +468,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(singleLayerHamtMultiBlockFilesFixture.MustGetCid()).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								singleLayerHamtMultiBlockFilesFixture.MustGetCid(),
@@ -507,8 +492,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
@@ -532,8 +516,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
@@ -557,8 +540,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
@@ -582,8 +564,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
@@ -607,8 +588,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),
@@ -632,8 +612,7 @@ func TestTrustlessCarEntityBytes(t *testing.T) {
 				Status(200).
 				Body(
 					IsCar().
-						HasRoot(subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt")).
-						MightHaveNoRoots().
+						IgnoreRoots().
 						HasBlocks(
 							flattenStrings(t,
 								subdirWithMixedBlockFiles.MustGetCid("subdir", "multiblock.txt"),